ollama celery task abstracted to its own module
